pyproject.toml
[project]
name = "youtube-summarizer"
version = "0.2.0"
description = "Download and summarize YouTube video transcripts using AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "youtube-transcript-api>=0.6.0",
    "pydantic-ai-slim>=0.2.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40.0"]
openai = ["openai>=1.0.0"]
gemini = ["google-generativeai>=0.8.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
]

[project.scripts]
youtube-summarizer = "youtube_summarizer.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
addopts = "-m 'not integration'"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests as integration tests",
]
